freebsd port
There are commands that only work on Linux; rename those files.

There are a few places that int style casts are needed.
