+++ cmds/core/dmesg/dmesg_linux.go
@@ -2,9 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-//go:build !plan9 && !windows
-
-// dmesg reads the system log.
+// dmesg reads the Linux system log.
 //
 // Synopsis:
 //
