+++ cmds/core/dmesg/dmesg_linux_test.go
